🎨 update

+++ fish4j-api/src/main/java/io/github/danniod/fish4j/entites/chatroom/RedPacketStatusMessage.java
@@ -5,13 +5,11 @@ import io.github.danniod.fish4j.entites.ChatroomMessage;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
-import java.math.BigInteger;
-
 @EqualsAndHashCode(callSuper = true)
 public @Data class RedPacketStatusMessage extends ChatroomMessage {
 
     @JsonProperty("oId")
-    private BigInteger id;
+    private Long id;
 
     private Integer count;
